MMFMWK-9110 dsp_wrapper: fix get wrong value from dsp
long type is 64bit in arm platform and dsp returned 32bit value.
Can't get the returned val in long type. Or the val is not correct.

(cherry picked from commit 0bd6b563bf4a5cde354be3bc631094c685829da4)
(cherry picked from commit 4426f159f0c9c3b1d13baedfa66710d89bf35d8e)

+++ dsp_wrapper/src/dsp_dec.c
@@ -298,8 +298,8 @@ int comp_process(UniACodec_Handle pua_handle,
 		return ACODEC_ERROR_STREAM;
 	} else if (error > XAF_NO_ERR) {
 		/* Get event from dsp */
-		long *config_buf = (long *)comp_info[0];
-		long decode_err = *config_buf;
+		int *config_buf = (int *)comp_info[0];
+		int decode_err = *config_buf;
 		free(config_buf);
 		/* get actual decode err */
 		if (decode_err & XA_FATAL_ERROR)
@@ -374,8 +374,8 @@ int comp_flush_msg(UniACodec_Handle pua_handle)
 			return ACODEC_ERROR_STREAM;
 		} else if (error > XAF_NO_ERR) {
 			/* Get event from dsp */
-			long *config_buf = (long *)comp_info[0];
-			long decode_err = *config_buf;
+			int *config_buf = (int *)comp_info[0];
+			int decode_err = *config_buf;
 			fprintf(stderr, "flush, get event, %d\n", decode_err);
 			free(config_buf);
 			continue;
